requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
minio == 7.1.14
apache-airflow-providers-docker == 3.7.1
apache-airflow-providers-slack == 8.1.0
